# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "matcheck Mobile API",
    "version": "1.0.0",
    "description": "API для внешних клиентов matcheck (мобильные приложения, интеграции).\n\nБазовый URL: `https://matcheck.fvds.ru/api/v1/`.\n\nAuth: Bearer JWT (Ed25519). Для мобильных клиентов используется заголовок `X-Client-Type: mobile` — в этом случае refresh-token возвращается в теле ответа `/auth/login` и `/auth/refresh`, cookies не устанавливаются.\n\nПодробная документация: см. `docs/MOBILE_API.md` в репозитории.",
    "contact": {
      "name": "matcheck",
      "url": "https://matcheck.fvds.ru"
    }
  },
  "servers": [
    {
      "url": "https://matcheck.fvds.ru",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519 JWT access-token, полученный из POST /auth/login или /auth/refresh. TTL 900 сек, обновляется через /auth/refresh."
      }
    },
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "inspector_kpp"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "isActive",
          "createdAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "manager",
                  "inspector_kpp"
                ]
              },
              "isActive": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "role",
              "isActive",
              "createdAt"
            ]
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshExpiresIn": {
            "type": "number"
          }
        },
        "required": [
          "accessToken",
          "expiresIn",
          "user"
        ]
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshExpiresIn": {
            "type": "number"
          }
        },
        "required": [
          "accessToken",
          "expiresIn"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход по email/паролю",
        "description": "Возвращает access-token (Ed25519 JWT, TTL 900 сек). Для мобильных клиентов с заголовком `X-Client-Type: mobile` дополнительно возвращает `refreshToken` и `refreshExpiresIn` в теле ответа; для веб-клиентов refresh-token устанавливается в HttpOnly-cookie `__Host-refresh`.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "mobile"
              ],
              "description": "Установить значение `mobile` для нативных клиентов. В этом случае refresh-token возвращается в теле ответа, а cookies не устанавливаются."
            },
            "required": false,
            "name": "x-client-type",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учётные данные или аккаунт неактивен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "Аккаунт временно заблокирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Превышен rate-limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Обновить access-token",
        "description": "Принимает refresh-token из HttpOnly-cookie `__Host-refresh` (веб) либо из заголовка `Authorization: Bearer <refreshToken>` при `X-Client-Type: mobile`. Реализует rotation: старый refresh-token инвалидируется, выдаётся новый. Reuse уже использованного refresh → 401 + полная инвалидация сессии.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "mobile"
              ],
              "description": "Установить значение `mobile` для нативных клиентов. В этом случае refresh-token возвращается в теле ответа, а cookies не устанавливаются."
            },
            "required": false,
            "name": "x-client-type",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Новый access-token (+ refresh-token в теле для mobile)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh-token отсутствует, истёк или инвалидирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Завершить сессию",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Сессия инвалидирована",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Не авторизовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Профиль текущего пользователя",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}